Assignment v1.2
The third commit. I decided on a scale of 1 pixel to 2 meters. This way I will be able to have a maximum distance across the screen of 1600m. This very easily fits the calculated maximum range of the simulation of 920m including the extra space the wookie is offset from the right.  

Whats done so far:

--phsyics doc
--start screen
--sprites
--new objects
--background
--scale
--starting values

What to do for next commit

--add in actual simulation for detonating
--allow for activation a player input