Assignment v1.3
In this commit we have finally added the simulation part to the project. To start the game all you need to do is press space. The ball will fly then land on the ground. Give it a sec and it will also ake a nice boom(no sound right now).

Done so far
--physics doc
--start screen
--assets
--new obects
--background
--starting values
--starting positions
--scale
--physics simulation
--little boom animation
--fully able to be reset by repressing input button

To do for next commit
--add ability to change starting values and adjust simulation