Assignment v1.4
This commit adds in two steps, I was originally gonna just do the labels then do a separate commit with the changing but since the template came with ImGui the first part was really quick. That means this commit has both the labels and the controlling of labels. It is a fairly big change. It adds labels, the ability to change the labels mid test to make your own sims and 4 buttons, 1 reset, and 3 solve buttons which solve for angle, range, and max range.

Things done
--physics doc
--start screen
--assets
--new objects
--background
--scale
--initial values
--initial positions
--physics sim
--boom effect
--reset
--labels
--adjust labels in game
--reset button
--solve for angle (based on range and vel)
--solve for range (based on angle and vel)
--solve for max range (based on vel)

Things to do for next commit

--nothing
--except video demonstration